Refactor BeatEngine: Make it a GameObject
Refactor BeatEngine so that it can be attached to a GameObject and
send events to BeatActors in the scene.
This way, less computations are made when calling utility functions
and higher precision time such as dspTime can be used. A prefab is
also created from the newly added GameObject for easier use.

Now, BeatEngine will send the BeatEvent when detecting a beat
and any script that inherits from the newly added interface IBeatActor
and that registered its OnBeat function to the BeatEngine can react.
This greatly simplifies the logic that was before in Update where every
GameObject had to have its own logic for detecting the beat event.