Refactor BeatPatternResolver to be more friendly
Refactor revolves around the fact that now a BeatPatternButton component
needs to be added to a GameObject along with a BoxCollider in order for it
to receive events when some beat patterns are resolved. All the code inside
Update methods in objects such as DefenseSpawner, DefenseController or
ResourcesMineController has been moved inside functions that are called
upon event. The whole update of the resolution logic is done inside
BeatPatternButton. Prefabs wer modified to include the BeatPatternButton script