Add level handling for Upgrades
[Upgrade]
- Force constructor taking parameters to construct upgrade correctly
- Add a AcquireLevel function which increases the level towards a
max level and updates the price
- Add a virtual function ComputePriceFromLevel to update price based
on level

[Defense*Upgrade]
- Update construction to reflect changes in Upgrade

[DefensePowerUpgrade & DefenseRangeUpgrade]
- Implement ComputePriceFromLevel

[UpgradeButton]
- Add levelText gameobject to the UpgradeButton prefab
- Pass levelText to UpgradeButtonController in order to display level

Also refactor the way UpgradesController and ShopController communicate.
Now, an event UpgradeModified is launched by UpgradesController and
ShopController acts upon it, updating its graphics (UpgradeButtonController)
Also moved resources handling to controller, so that controller decides
whether or not we are able to buy an upgrade.
Add an exception thrown when an Upgrade was already acquired before