Suppress warning CS0649
It is a known issue that when using SerializeField on private memebers
to expose them through the Unity inspector, a warning is thrown because
the variable is never assigned a value, even though it is, only through
the editor.

In the case of SkillPointsCounter, surround the variable definition with
pragma instructions to disable and then restore the warning.