Implement player save
Add a Player class which is reponsible for serializing and deserializing
to and from a yaml file containing the player's progress. So far it
includes: the number of skill points, the upgrades bought and the levels
beaten.

The classes LevelController and SkillPointsController were removed as
they would have duplicated logic with the new Player class.

[Upgrade]
- Add an Init function from a level so that the deserialized upgrade
level from the player's upgrades can be used to initialize an Upgrade
[UpgradesController]
- Use the newly added Init function from Upgrade to initialize all
bought upgrades to the level acquired by the player

[IntroScreenController]
- Either load or create the default player data on disk in Start method

[LevelData]
- Modify so that it can be used as serialized data. All the unnecessary
data was removed and only the achievements boolean were kept which are
serialized in the player's data

[LevelSelectController]
- Since the LevelData was modified to be used solely through the player,
modify the selectedLevel property to be the selectedLevelName

[LevelSelectButton]
- Use player's deserialized data to initialize the button's state

[LevelWonSceneController]
- Modify logic so that based on whether or not the player already beat
the level or obtained achievements skill points are not granted a second
time. If it was the first time, use Player's function to serialize the
fact that the level was beaten and whether or not achievements were
obtained

[Other files]
- All other changes reflect the fact that Player replaced logic from
SkillPointsController and LevelController